Add SuperShape output and enhance curve processing
Added a new output parameter "SuperShape" to the `RegisterOutputParams` method to output a list of points. Introduced `pData_points` to store points related to the SuperShape. Changed the Z-axis vector for plane creation from `Vector3d.ZAxis` to `(0, 0, -1)`. Added logic to retrieve the operation UI and indices for program and operation. Enhanced handling of extrusion operation's action states and activation mode.

Modified the creation of `SuperShape` objects to use the `planes` list instead of `pathPlanes`. Introduced `processedCurves` to handle curves based on their length. Added logic to process each curve in `splitCurves`, checking their length and performing different operations based on a threshold of 10 units. Created `SMTPData` objects for each curve and added them to the `smtPlugin.UserData` dictionary. Added the origin of `SMTPData` points to the `pData_points` list. Updated the `SuperShape` creation to use the new `SMTPData` and added them to the `shapes` array. Set the output data for the new "SuperShape" parameter in the `DA.SetDataList` method.